rename bytes/bytes.ts to bytes/mod.ts
