+++ bytes/mod.ts
